MNIST/requirements.txt
scikit-learn
numpy
Pillow
joblib
flask
werkzeug
